Avoid to destroy category object when it has work objects, and Make better form using persisted method

+++ app/controllers/categories_controller.rb
@@ -38,8 +38,12 @@ class CategoriesController < ApplicationController
   
   def destroy
     category = Category.find params[:id]
-    category.destroy!
-    redirect_to categories_path
+    if category.works.presence == nil
+      category.destroy!
+      redirect_to categories_path
+    else
+      render `index`
+    end
   end
 
   private
